implement verify jwt :technologist:
