+++ routes/userRouter.js
@@ -1,7 +1,11 @@
 var express = require("express");
 var router = express.Router();
+
 const userController = require("../controllers/userController");
+// const verifyToken = require("../middlewares/verifyJWT");
+
 router.use(express.json())
+// router.use(verifyToken)
 
 router.post('/signin', userController.createUser);
